+++ packages/astro/CHANGELOG.md
@@ -1,3 +1,11 @@
+## @lunora/astro [1.0.0-alpha.108](https://github.com/anolilab/lunora/compare/@lunora/astro@1.0.0-alpha.107...@lunora/astro@1.0.0-alpha.108) (2026-09-02)
+
+
+### Dependencies
+
+* **@lunora/client:** upgraded to 1.0.0-alpha.71
+* **@lunora/runtime:** upgraded to 1.0.0-alpha.88
+
 ## @lunora/astro [1.0.0-alpha.107](https://github.com/anolilab/lunora/compare/@lunora/astro@1.0.0-alpha.106...@lunora/astro@1.0.0-alpha.107) (2026-09-01)