+++ packages/notify/CHANGELOG.md
@@ -1,3 +1,10 @@
+## @lunora/notify [1.0.0-alpha.33](https://github.com/anolilab/lunora/compare/@lunora/notify@1.0.0-alpha.32...@lunora/notify@1.0.0-alpha.33) (2026-09-02)
+
+
+### Dependencies
+
+* **@lunora/errors:** upgraded to 1.0.0-alpha.29
+
 ## @lunora/notify [1.0.0-alpha.32](https://github.com/anolilab/lunora/compare/@lunora/notify@1.0.0-alpha.31...@lunora/notify@1.0.0-alpha.32) (2026-09-01)