+++ packages/react-native/CHANGELOG.md
@@ -1,3 +1,11 @@
+## @lunora/react-native [1.0.0-alpha.51](https://github.com/anolilab/lunora/compare/@lunora/react-native@1.0.0-alpha.50...@lunora/react-native@1.0.0-alpha.51) (2026-09-02)
+
+
+### Dependencies
+
+* **@lunora/client:** upgraded to 1.0.0-alpha.71
+* **@lunora/react:** upgraded to 1.0.0-alpha.76
+
 ## @lunora/react-native [1.0.0-alpha.50](https://github.com/anolilab/lunora/compare/@lunora/react-native@1.0.0-alpha.49...@lunora/react-native@1.0.0-alpha.50) (2026-09-01)