+++ packages/react/CHANGELOG.md
@@ -1,3 +1,12 @@
+## @lunora/react [1.0.0-alpha.76](https://github.com/anolilab/lunora/compare/@lunora/react@1.0.0-alpha.75...@lunora/react@1.0.0-alpha.76) (2026-09-02)
+
+
+### Dependencies
+
+* **@lunora/client:** upgraded to 1.0.0-alpha.71
+* **@lunora/errors:** upgraded to 1.0.0-alpha.29
+* **@lunora/ratelimit:** upgraded to 1.0.0-alpha.35
+
 ## @lunora/react [1.0.0-alpha.75](https://github.com/anolilab/lunora/compare/@lunora/react@1.0.0-alpha.74...@lunora/react@1.0.0-alpha.75) (2026-09-01)