+++ packages/storage/CHANGELOG.md
@@ -1,3 +1,11 @@
+## @lunora/storage [1.0.0-alpha.47](https://github.com/anolilab/lunora/compare/@lunora/storage@1.0.0-alpha.46...@lunora/storage@1.0.0-alpha.47) (2026-09-02)
+
+
+### Dependencies
+
+* **@lunora/errors:** upgraded to 1.0.0-alpha.29
+* **@lunora/platform:** upgraded to 1.0.0-alpha.24
+
 ## @lunora/storage [1.0.0-alpha.46](https://github.com/anolilab/lunora/compare/@lunora/storage@1.0.0-alpha.45...@lunora/storage@1.0.0-alpha.46) (2026-09-01)